+++ src/Options.hpp
@@ -1,0 +1,8 @@
+#ifndef OPTIONS_H
+#define OPTIONS_H
+
+struct Options {
+	bool formattedIntAsUserData = false;
+	bool temporalTypesAsUserData = true;
+};
+#endif /* OPTIONS_H */