+++ CHANGELOG.md
@@ -5,6 +5,76 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0](https://github.com/LebJe/toml.lua/releases/tag/0.2.0) - 2023-02-12
+
+### Added
+
+-   Updated to toml++ v3.3.0
+-   Added `terseKeyValuePairs` to the list of formatting options for `toml.encode`
+-   Tables can be made inline.
+-   `toml.decode` can decode date, time and date-time into userdata (as before) or plain tables:
+
+    ```lua
+    local tomlStr = [[
+    date = 1979-05-27
+    time = 07:32:00
+    datetime = 1979-05-27T07:32:00-07:00
+    ]]
+
+    local table1 = toml.decode(tomlStr, { temporalTypesAsUserData = true })
+    local table2 = toml.decode(tomlStr, { temporalTypesAsUserData = false })
+
+    print(inspect(table1))
+    --[[
+    {
+    	date = <userdata 1> -- 1979-05-27,
+    	time = <userdata 2> -- 07:32:00,
+    	datetime = <userdata 3> -- 1979-05-27T07:32:00-07:00
+    }
+    --]]
+
+    print(inspect(table2))
+    --[[
+    {
+    	date = {
+    		day = 27,
+    		month = 5,
+    		year = 1979
+    	},
+    	datetime = {
+    		date = {
+    			day = 27,
+    			month = 5,
+    		year = 1979
+    		},
+    		time = {
+    			hour = 7,
+    			minute = 32,
+    			nanoSecond = 0,
+    			second = 0
+    		},
+    		timeOffset = { minutes = -420 }
+    	},
+    	time = {
+    		hour = 7,
+    		minute = 32,
+    		nanoSecond = 0,
+    		second = 0
+    	}
+    }
+    --]]
+    ```
+
+-   Test suite now runs on Windows
+
+### Changed
+
+-   Use luaunit instead of busted for testing
+
+### Fixed
+
+-   Boolean values are decoded as booleans instead of integers. (#6)
+
 ## [0.1.1](https://github.com/LebJe/toml.lua/releases/tag/0.1.1) - 2022-06-14
 
 ### Added